C58 compiles.

+++ chapter-58-itemstats/src/ai/approach_ai_system.rs
@@ -33,8 +33,8 @@ impl<'a> System<'a> for ApproachAI {
             if path.success && path.steps.len()>1 {
                 let mut idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = false;
-                pos.x = path.steps[1] % map.width;
-                pos.y = path.steps[1] / map.width;
+                pos.x = path.steps[1] as i32 % map.width;
+                pos.y = path.steps[1] as i32 / map.width;
                 entity_moved.insert(entity, EntityMoved{}).expect("Unable to insert marker");
                 idx = map.xy_idx(pos.x, pos.y);
                 map.blocked[idx] = true;
